Implemetned Colin Faulkingham's ImageExport class. Made a few modifications so that it exports a dbi object. Image open, drawing open and drawing save are now implimented. Still need to add code to update the rest of the GUI controls on image/drawing load.

+++ src/com/tautic/drawbotcontrol/DrawBotImage.java
@@ -12,7 +12,7 @@ public class DrawBotImage implements Serializable {
 	public int paperHeight;
 	public char drawingSpeed;
 	public char drawingDelay; 
-	public ArrayList<Byte> imageData; //Contains image pixel data.
+	public ArrayList<Integer> imageData; //Contains image pixel data.
 	
 		public void setColumns(char MSB, char LSB) {
 			columns = MSB >> 8 | LSB;
